files cleaned, lint satisfied.
