--- examples/cpp/beesandflowers/flowers_log.csv
@@ -1,401 +1,0 @@
-id,x,y,nectar
-1,0,0,24.3658
-6,0,5,35.8717
-11,0,10,35.9001
-16,0,15,10.7579
-21,0,20,37.3321
-26,0,25,49.5371
-31,0,30,35.9438
-36,0,35,42.212
-41,0,40,46.1942
-46,0,45,27.0852
-51,0,50,48.4241
-56,0,55,23.0335
-61,0,60,21.0893
-66,0,65,11.7221
-71,0,70,16.1631
-76,0,75,16.1068
-81,0,80,31.2847
-86,0,85,20.1925
-91,0,90,38.7841
-96,0,95,36.5171
-501,5,0,22.8213
-506,5,5,38.4367
-511,5,10,11.6686
-516,5,15,28.1005
-521,5,20,20.9745
-526,5,25,47.1759
-531,5,30,41.7979
-536,5,35,43.8008
-541,5,40,41.5657
-546,5,45,49.0516
-551,5,50,49.8962
-556,5,55,24.694
-561,5,60,44.8275
-566,5,65,38.251
-571,5,70,20.8663
-576,5,75,13.6848
-581,5,80,47.0261
-586,5,85,42.4271
-591,5,90,26.546
-596,5,95,18.3368
-1001,10,0,46.9505
-1006,10,5,39.2443
-1011,10,10,39.2124
-1016,10,15,30.1553
-1021,10,20,46.4993
-1026,10,25,38.7231
-1031,10,30,48.0947
-1036,10,35,26.3862
-1041,10,40,20.9435
-1046,10,45,11.7529
-1051,10,50,33.8525
-1056,10,55,22.316
-1061,10,60,48.0909
-1066,10,65,22.1804
-1071,10,70,44.8681
-1076,10,75,35.6176
-1081,10,80,27.8044
-1086,10,85,46.9102
-1091,10,90,16.1903
-1096,10,95,25.9288
-1501,15,0,26.2944
-1506,15,5,43.5629
-1511,15,10,14.0103
-1516,15,15,31.0146
-1521,15,20,32.4707
-1526,15,25,44.7412
-1531,15,30,35.6243
-1536,15,35,10.8516
-1541,15,40,46.7302
-1546,15,45,32.1799
-1551,15,50,48.81
-1556,15,55,18.3909
-1561,15,60,45.983
-1566,15,65,49.4055
-1571,15,70,34.915
-1576,15,75,20.4425
-1581,15,80,36.8974
-1586,15,85,11.3033
-1591,15,90,28.4377
-1596,15,95,45.1995
-2001,20,0,38.777
-2006,20,5,42.6688
-2011,20,10,14.3954
-2016,20,15,33.6658
-2021,20,20,13.216
-2026,20,25,15.6887
-2031,20,30,17.7646
-2036,20,35,36.9391
-2041,20,40,45.338
-2046,20,45,39.424
-2051,20,50,15.847
-2056,20,55,28.4718
-2061,20,60,31.1931
-2066,20,65,29.0256
-2071,20,70,10.5372
-2076,20,75,12.5691
-2081,20,80,38.0319
-2086,20,85,27.7429
-2091,20,90,32.402
-2096,20,95,19.9528
-2501,25,0,39.4082
-2506,25,5,13.3268
-2511,25,10,18.568
-2516,25,15,42.8596
-2521,25,20,37.8042
-2526,25,25,18.5815
-2531,25,30,49.3816
-2536,25,35,18.2709
-2541,25,40,44.1497
-2546,25,45,20.5645
-2551,25,50,10.9152
-2556,25,55,45.8167
-2561,25,60,41.2502
-2566,25,65,32.2952
-2571,25,70,28.7335
-2576,25,75,16.5843
-2581,25,80,39.2795
-2586,25,85,26.1422
-2591,25,90,42.3913
-2596,25,95,45.6682
-3001,30,0,45.8026
-3006,30,5,29.7911
-3011,30,10,45.5649
-3016,30,15,41.1182
-3021,30,20,29.385
-3026,30,25,46.099
-3031,30,30,12.1469
-3036,30,35,33.1366
-3041,30,40,27.3239
-3046,30,45,31.3204
-3051,30,50,37.5872
-3056,30,55,34.9071
-3061,30,60,24.004
-3066,30,65,42.064
-3071,30,70,10.8745
-3076,30,75,14.056
-3081,30,80,44.9444
-3086,30,85,19.0173
-3091,30,90,13.1407
-3096,30,95,15.1487
-3501,35,0,28.7757
-3506,35,5,39.3252
-3511,35,10,17.2018
-3516,35,15,29.3384
-3521,35,20,25.2599
-3526,35,25,23.9277
-3531,35,30,41.1498
-3536,35,35,43.5743
-3541,35,40,48.9576
-3546,35,45,34.3168
-3551,35,50,31.4978
-3556,35,55,25.4894
-3561,35,60,47.8001
-3566,35,65,10.8292
-3571,35,70,19.0511
-3576,35,75,41.357
-3581,35,80,33.8471
-3586,35,85,16.5512
-3591,35,90,13.8488
-3596,35,95,17.7127
-4001,40,0,36.9625
-4006,40,5,28.8211
-4011,40,10,16.2739
-4016,40,15,20.0722
-4021,40,20,14.634
-4026,40,25,39.6142
-4031,40,30,13.2106
-4036,40,35,12.7703
-4041,40,40,22.6394
-4046,40,45,20.2284
-4051,40,50,18.7299
-4056,40,55,19.3254
-4061,40,60,36.922
-4066,40,65,25.7077
-4071,40,70,34.0563
-4076,40,75,42.43
-4081,40,80,17.5022
-4086,40,85,45.0099
-4091,40,90,19.6961
-4096,40,95,46.7827
-4501,45,0,21.4801
-4506,45,5,20.3232
-4511,45,10,22.1067
-4516,45,15,30.9119
-4521,45,20,44.4835
-4526,45,25,20.8125
-4531,45,30,11.1795
-4536,45,35,13.5761
-4541,45,40,10.2319
-4546,45,45,39.6539
-4551,45,50,45.1337
-4556,45,55,22.9831
-4561,45,60,26.4604
-4566,45,65,32.3839
-4571,45,70,24.045
-4576,45,75,27.0158
-4581,45,80,20.6021
-4586,45,85,45.1349
-4591,45,90,24.8681
-4596,45,95,43.6931
-5001,50,0,37.5621
-5006,50,5,10.5884
-5011,50,10,30.4773
-5016,50,15,43.5052
-5021,50,20,37.3025
-5026,50,25,46.3936
-5031,50,30,42.1267
-5036,50,35,24.8631
-5041,50,40,43.631
-5046,50,45,22.6127
-5051,50,50,24.3694
-5056,50,55,46.319
-5061,50,60,47.9328
-5066,50,65,25.1503
-5071,50,70,28.614
-5076,50,75,25.0241
-5081,50,80,48.7213
-5086,50,85,46.251
-5091,50,90,47.6776
-5096,50,95,31.8099
-5501,55,0,45.4626
-5506,55,5,41.8334
-5511,55,10,36.4803
-5516,55,15,27.7069
-5521,55,20,40.2593
-5526,55,25,46.3946
-5531,55,30,20.5077
-5536,55,35,11.9265
-5541,55,40,19.5881
-5546,55,45,33.4215
-5551,55,50,28.8389
-5556,55,55,25.8969
-5561,55,60,10.4196
-5566,55,65,29.1585
-5571,55,70,28.0467
-5576,55,75,36.7596
-5581,55,80,22.714
-5586,55,85,47.0404
-5591,55,90,48.9131
-5596,55,95,12.9902
-6001,60,0,32.9984
-6006,60,5,45.8617
-6011,60,10,28.8644
-6016,60,15,24.8188
-6021,60,20,26.9227
-6026,60,25,27.3354
-6031,60,30,13.7609
-6036,60,35,49.4636
-6041,60,40,45.4167
-6046,60,45,24.4415
-6051,60,50,27.6088
-6056,60,55,45.3547
-6061,60,60,12.8242
-6066,60,65,37.4034
-6071,60,70,44.8674
-6076,60,75,16.6243
-6081,60,80,10.2513
-6086,60,85,33.5738
-6091,60,90,10.3151
-6096,60,95,38.225
-6501,65,0,48.8664
-6506,65,5,22.2014
-6511,65,10,41.6745
-6516,65,15,13.2931
-6521,65,20,18.1137
-6526,65,25,25.6582
-6531,65,30,42.1673
-6536,65,35,26.8162
-6541,65,40,20.1184
-6546,65,45,20.402
-6551,65,50,32.1559
-6556,65,55,10.1321
-6561,65,60,41.174
-6566,65,65,23.1239
-6571,65,70,45.9826
-6576,65,75,14.7353
-6581,65,80,22.3784
-6586,65,85,16.8391
-6591,65,90,23.7452
-6596,65,95,11.2556
-7001,70,0,32.2347
-7006,70,5,38.3539
-7011,70,10,24.7187
-7016,70,15,48.7938
-7021,70,20,24.6182
-7026,70,25,42.9988
-7031,70,30,21.6453
-7036,70,35,32.4648
-7041,70,40,39.6258
-7046,70,45,25.2827
-7051,70,50,35.1197
-7056,70,55,39.8254
-7061,70,60,49.4719
-7066,70,65,23.7163
-7071,70,70,48.4168
-7076,70,75,40.0722
-7081,70,80,14.3061
-7086,70,85,44.5452
-7091,70,90,26.6436
-7096,70,95,38.9191
-7501,75,0,12.6621
-7506,75,5,32.5406
-7511,75,10,13.3094
-7516,75,15,10.302
-7521,75,20,26.1715
-7526,75,25,48.8336
-7531,75,30,38.9679
-7536,75,35,30.4844
-7541,75,40,21.3099
-7546,75,45,46.371
-7551,75,50,16.4795
-7556,75,55,29.7907
-7561,75,60,23.2726
-7566,75,65,38.4863
-7571,75,70,25.9188
-7576,75,75,33.1228
-7581,75,80,27.5126
-7586,75,85,20.5139
-7591,75,90,15.3581
-7596,75,95,40.6442
-8001,80,0,24.0812
-8006,80,5,21.7105
-8011,80,10,15.3027
-8016,80,15,35.854
-8021,80,20,29.8325
-8026,80,25,20.2504
-8031,80,30,14.798
-8036,80,35,10.3784
-8041,80,40,29.2962
-8046,80,45,25.8491
-8051,80,50,47.3288
-8056,80,55,41.643
-8061,80,60,44.3761
-8066,80,65,27.1764
-8071,80,70,40.8607
-8076,80,75,18.7796
-8081,80,80,11.994
-8086,80,85,28.7365
-8091,80,90,15.232
-8096,80,95,28.8704
-8501,85,0,19.9787
-8506,85,5,23.0439
-8511,85,10,42.4189
-8516,85,15,49.6174
-8521,85,20,12.7246
-8526,85,25,45.7428
-8531,85,30,41.9248
-8536,85,35,19.1639
-8541,85,40,42.7749
-8546,85,45,19.9467
-8551,85,50,38.2096
-8556,85,55,11.7117
-8561,85,60,48.29
-8566,85,65,27.7903
-8571,85,70,17.4615
-8576,85,75,26.8388
-8581,85,80,49.537
-8586,85,85,36.853
-8591,85,90,46.8866
-8596,85,95,22.0073
-9001,90,0,25.707
-9006,90,5,15.4338
-9011,90,10,17.441
-9016,90,15,26.6732
-9021,90,20,37.1267
-9026,90,25,21.7836
-9031,90,30,25.0609
-9036,90,35,30.0073
-9041,90,40,12.6304
-9046,90,45,31.1831
-9051,90,50,45.9095
-9056,90,55,36.0021
-9061,90,60,48.295
-9066,90,65,27.1316
-9071,90,70,35.2749
-9076,90,75,45.8489
-9081,90,80,26.5912
-9086,90,85,22.1382
-9091,90,90,14.3508
-9096,90,95,31.5936
-9501,95,0,11.2994
-9506,95,5,34.7597
-9511,95,10,21.5916
-9516,95,15,34.0532
-9521,95,20,24.0706
-9526,95,25,32.3734
-9531,95,30,29.6428
-9536,95,35,46.2754
-9541,95,40,31.2519
-9546,95,45,26.8134
-9551,95,50,20.9946
-9556,95,55,45.6577
-9561,95,60,40.6703
-9566,95,65,23.237
-9571,95,70,29.6241
-9576,95,75,31.8105
-9581,95,80,31.1723
-9586,95,85,12.1477
-9591,95,90,22.8725
-9596,95,95,10.321
