+++ src/CMakeLists.txt
@@ -300,7 +300,7 @@ SET(SRC_INCLUDE
     ${FLAMEGPU_ROOT}/include/flamegpu/detail/TestSuiteTelemetry.h
     ${FLAMEGPU_ROOT}/include/flamegpu/detail/JitifyCache.h
     ${FLAMEGPU_ROOT}/include/flamegpu/stockAgent/subModels/AbstractSubmodels.h
-    ${FLAMEGPU_ROOT}/include/flamegpu/stockAgent/subModels/AddOneAgent.h
+    ${FLAMEGPU_ROOT}/include/flamegpu/stockAgent/subModels/SingleAgentDiscreteMovement.h
 )
 SET(SRC_FLAMEGPU
     ${FLAMEGPU_ROOT}/src/flamegpu/exception/FLAMEGPUException.cpp
@@ -387,7 +387,7 @@ SET(SRC_FLAMEGPU
     ${FLAMEGPU_ROOT}/src/flamegpu/detail/wddm.cu
     ${FLAMEGPU_ROOT}/src/flamegpu/detail/JitifyCache.cu
     ${FLAMEGPU_ROOT}/src/flamegpu/detail/TestSuiteTelemetry.cpp
-    ${FLAMEGPU_ROOT}/src/flamegpu/stockAgent/subModels/AddOneAgent.cu
+    ${FLAMEGPU_ROOT}/src/flamegpu/stockAgent/subModels/SingleAgentDiscreteMovement.cu
 )
 SET(SRC_DYNAMIC
     ${DYNAMIC_VERSION_SRC_DEST}
